+++ tests/tel2puml/otel_to_pv/otel_ingestion/conftest.py
@@ -4,7 +4,7 @@ import pytest
 from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
     FieldSpec,
 )
-from tel2puml.otel_to_pv.otel_ingestion.json_jq_converter import (
+from tel2puml.otel_to_pv.data_sources.json_data_source.json_jq_converter import (
     JQVariableTree,
 )
